# backend/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is a sample server api for beacon based attendance system"
  version: "1.0.0"
  title: "Swagger beacon-based attendance system"
  contact:
    email: "mdsazzadul.islam@tuni.fi"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "127.0.0.1:8080"
basePath: "/v1"
paths:
...
  /room:
    post:
      tags:
      - "room"
      summary: "adding a room"
      description: "Add a new Room"
      operationId: "addRoom"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "room object that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/roomBody"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid input"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Room"
    get:
      tags:
      - "room"
      summary: "Get all room"
      description: "Get all room from database"
      operationId: "getAllRooms"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/roomList"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Room"
  /room/{roomId}:
    get:
      tags:
      - "room"
      summary: "Find an room by its ID"
      description: "IDs must be positive integers"
      operationId: "getRoomById"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "roomId"
        description: "ID of room that needs to be fetched"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/roomItem"
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Room"
    put:
      tags:
      - "room"
      summary: "Update an room by its ID"
      description: "IDs must be positive integers"
      operationId: "updateRoom"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "roomId"
        description: "ID of the room"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
      - in: "body"
        name: "body"
        description: "room object that needs to be added to the database"
        required: true
        schema:
          $ref: "#/definitions/roomBody"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid input"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Room"
    delete:
      tags:
      - "room"
      summary: "Delete a room"
      description: "Delete a room from database for path roomId"
      operationId: "deleteRoom"
      produces:
      - "application/json"
      parameters:
      - in: "header"
        name: "api_key"
        required: true
        type: "string"
      - in: "path"
        name: "roomId"
        required: true
        type: "integer"
        minimum: 1
        format: "int32"
        description: "ID of the room"
      responses:
        201:
          description: "successful operation"
          schema:
            $ref: '#/definitions/responseWithMessage'
        400:
          description: "Invalid ID supplied"
        401:
          description: "Authorization failed"
          schema:
            $ref: '#/definitions/responseWithMessage'
        404:
          description: "The resource you are looking is not found."
          schema:
            $ref: '#/definitions/404'
      x-swagger-router-controller: "Room"
...
definitions:
  '404':
    type: object
    additionalProperties: false
    properties:
      error:
        type: string
    required:
      - error
  responseWithMessage:
    type: object
    additionalProperties: false
    properties:
      message:
        type: string
    required:
      - message
...
  roomBody:
    type: "object"
    additionalProperties: false
    properties:
      name:
        type: string
        minLength: 1
        example: "TC-110"
    required:
      - name
  roomItem:
    type: object
    additionalProperties: false
    properties:
      id:
        type: integer
        format: int32
      name:
        type: string
    required:
      - id
      - name
  roomList:
    type: array
    items:
      $ref: "#/definitions/roomItem"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "X-API-KEY"
    in: "header"